prom: Fix leak when a call fails with the prometheus endpoint
Some profiling data showed a goroutine leak. The reason is that
targetResults channel was never closed when the call to prometheus
endpoint fails. This is during the creation of prometheus
widgets in the console UI.

Also add support of request context for the call to prometheus endpoint
so it can get canceled when the caller context is canceled.